Stop treating MetaType as a proxy type
There's a lot of code checking `is_proxy_type` assuming that it's going
to find a literal type. That's not really true, and also the
`underlying()` method of a `MetaType` returns `Object` which is like
aspirationally true but not really useful and if we're depending on
this, it's probably a bug in disguise.