"Unable to resolve" → "resolves but is not imported"
There is now nothing catching the user if `CCC` is _only_ defined in a
`__package.rb` file. If the `__package.rb` file is empty, then we will
not report the "Unable to resolve constant" error that we otherwise
would have.

We might want to deal with that somehow. One way would be to abandon
this attempt to drop `<PackageSpecRegistry>`, but that is akin to
abandoning the approach of having `ClassOrModule` symbols own `Package`
symbols, which would mean we need a new way of integrating packages into
the symbol table.

Another way would be to somehow mark a symbol as failing to resolve if
the constant that it resolves to only has `__package.rb` locs. This
might work, because I think those locs should be there after namer and
before resolver, but it's worth thinking about. I don't think it would
catch when we resolve the constants in the `export` line of a
`__package.rb` though, because we haven't processed the
non-`__package.rb` files yet.

We could check after we've finished running `namer+resolver` for all the
packages to make sure that every `__package.rb` file has its namespace
defined somewhere. This fundamentally adds a join point after source
files back to checking a property of the __package.rb file, which is a
dependency that doesn't exist right now.